@@ -1,5 +1,5 @@
 {
-  "name": "solid-rtk",
+  "name": "react-rtk",
   "private": true,
   "version": "0.0.1",
   "type": "module",
@@ -13,11 +13,17 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "@solidjs/router": "^0.10.10",
-    "solid-js": "^1.8.7"
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.3"
   },
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@types/react": "^18.2.48",
+    "@types/react-dom": "^18.2.18",
+    "@types/react-router-dom": "^5.3.3",
+    "@vitejs/plugin-react-swc": "^3.5.0",
     "autoprefixer": "^10.4.17",
     "concurrently": "^8.2.2",
     "cssnano": "^6.0.3",
@@ -27,9 +33,8 @@
     "prettier": "^3.2.4",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.2.2",
-    "vite": "^5.0.8",
-    "vite-plugin-solid": "^2.8.0"
+    "typescript": "^5.3.3",
+    "vite": "^5.0.12"
   },
   "volta": {
     "node": "20.11.0"
